requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
numpy
python-telegram-bot==20.0b0
SQLAlchemy==2.0.0b4